Enhance user profile management and add new features
Updated IUserProfileRepository to use long for id in GetAsync.
Enhanced UserProfile class with new constructors, properties, and methods.
Added new package references for AutoMapper and oneof in HolaViaje.Social.csproj.
Updated Program.cs to add AddSocialFeatures method.
Added ServiceRegistration class in HolaViaje.Social.Features namespace.
Added IUserProfileApplication interface for user profile operations.
Added ServiceRegistration class in HolaViaje.Social.Features.Profiles namespace.
Implemented UserProfileApplication class for user profile management.
Added UserProfileErrorHelper class for error handling.
Added new model classes for user profile data representation.
Added new AutoMapper profile classes for domain and view model mapping.
Added new shared model classes for common data structures.
Added new AutoMapper profile classes for shared model mapping.

+++ src/HolaViaje.Social/Features/Profiles/Repository/IUserProfileRepository.cs
@@ -2,7 +2,7 @@
 
 public interface IUserProfileRepository
 {
-    Task<UserProfile?> GetAsync(int id, bool traking = false);
+    Task<UserProfile?> GetAsync(long id, bool traking = false);
     Task<UserProfile> CreateAsync(UserProfile userProfile);
     Task<UserProfile> UpdateAsync(UserProfile userProfile);
 }
